Cursor Changes and Drag from hit point
Need to test the cursor more, but the idea is to help the user avoid confusion about what tool they are using and if its active. The drag is just so you can rotate objects easily by grabing an outside spot. Could make a toggle to use the center or hit.point. possible ugly merge.